Update patches verified with T540p and USB vendor ID removal
Had a problem using this with my T540p. Diagnosed the issue down to having TWO loops run of the checker, one in PCI-mode and one in USB-mode. The PCI-mode check was patched but the USB-mode check was not patched (possibly missed with the first line I didn't attach a descriptive comment to - that didn't match bytes in my module version).